+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 ##### Enhancements
 
-* FunctionBodyLengthRule now does not count comment lines.  
+* TypeBodyLengthRule now does not count comment or whitespace lines.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#369](https://github.com/realm/SwiftLint/issues/369)
+
+* FunctionBodyLengthRule now does not count comment or whitespace lines.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#258](https://github.com/realm/SwiftLint/issues/258)